[MLIR] Create GPU utils library & move distribution utils
